pyproject.toml
[project]
name = "blasons"
version = "0.1.0"
description = "French Communes Coat of Arms Downloader"
requires-python = ">=3.10"
dependencies = [
    "polars>=1.0.0",
    "httpx>=0.27.0",
    "unidecode>=1.3.8",
    "rapidfuzz>=3.0.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
blasons = "blasons.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.black]
line-length = 88
target-version = "py310"
